Backend/requirements_bckup.txt
# FastAPI and Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database and ORM
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9

# Pydantic and Data Validation
pydantic==2.5.0
pydantic-settings==2.1.0
email-validator==2.1.1

# Authentication and Security
python-jose[cryptography]==3.3.0
PyJWT==2.8.0
passlib[bcrypt]==1.7.4

# Environment and Configuration
python-dotenv==1.0.0

# HTTP Client and External Services
requests==2.31.0
httpx==0.25.2

# Twilio for SMS/OTP
twilio==8.10.0

# AWS SDK for S3
boto3 ==1.28.60
aioboto3==12.0.0

# File Handling
aiofiles==23.2.1

# Date and Time
pytz==2023.3

# CORS is handled by FastAPI's CORSMiddleware (no separate package needed)

# Development and Testing (optional but recommended)
pytest==7.4.3
pytest-asyncio==0.21.1

# Additional utilities that might be needed
python-dateutil==2.8.2
typing-extensions==4.8.0

# For production deployment
gunicorn==21.2.0n
